+++ dev/config/src/vite/vite.test.config.ts
@@ -12,28 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import path from "node:path";
-import tsconfigPaths from "vite-tsconfig-paths";
 import { defineConfig } from "vitest/config";
-import VitePluginCloseAndCopy from "./vite-plugin-close-and-copy.js";
-import VitePluginSourcemapExclude from "./vite-plugin-sourcemap-exclude.js";
 
-export default function (tsConfigPath?: string) {
+export default function () {
 	const testTypeEnv = process.env.TEST_TYPE || "";
 	const testTypes = testTypeEnv.trim().split(",");
 	// @(|) globs any tests which don't have their type specified, e.g. myTest.test.ts. These are included even when filtering by test type because we don't know what type of test they are. Really, they should have their type specified.
 	// If we drop ^, there's a chance the tests with no type specified get ignored by accident, which we want to avoid. Ergo, include them by default.
 	const testTypeGlob = `@(|${testTypes.map((t) => (t ? `.${t.trim()}` : "")).join("|")})`;
 	console.log(`Filtering tests by type: ${testTypeGlob}`);
-
-	const plugins = [
-		VitePluginSourcemapExclude({ excludeNodeModules: true }),
-		VitePluginCloseAndCopy(),
-	];
-
-	if (tsConfigPath) {
-		plugins.push(tsconfigPaths({ projects: [path.resolve(tsConfigPath)] }));
-	}
 
 	return defineConfig({
 		build: {
@@ -71,6 +58,5 @@ export default function (tsConfigPath?: string) {
 			},
 			testTimeout: 10000,
 		},
-		plugins: plugins,
 	});
 }