+++ packages/api-express-router/src/middlewares/requestLoggerMiddleware.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { type LogLevel, getLogger, parseLogLevel } from "@prosopo/common";
+import { getLogger, parseLogLevel } from "@prosopo/common";
 import type { ProviderEnvironment } from "@prosopo/env";
 import type { NextFunction, Request, Response } from "express";
 import { v4 as uuidv4 } from "uuid";