+++ packages/locale/src/i18nFrontend.ts
@@ -18,7 +18,7 @@ import ChainedBackend from "i18next-chained-backend";
 import HttpBackend from "i18next-http-backend";
 import resourcesToBackend from "i18next-resources-to-backend";
 import { initReactI18next } from "react-i18next";
-import { LanguageSchema, Languages } from "./translations.js";
+import { i18nSharedOptions } from "./i18SharedOptions.js";
 
 const reactOptions: InitOptions = {
 	react: {
@@ -40,13 +40,18 @@ export function initializeI18n() {
 		i18n
 			// @ts-ignore
 			.use(ChainedBackend)
-			.use(I18nextBrowserLanguageDetector)
+			.use(
+				new I18nextBrowserLanguageDetector(null, {
+					order: ["cookie", "localStorage", "navigator"],
+					lookupQuerystring: "lng",
+					lookupCookie: "i18next",
+					lookupLocalStorage: "i18nextLng",
+					caches: ["localStorage", "cookie"],
+				}),
+			)
 			.use(initReactI18next)
 			.init({
-				debug: true,
-				fallbackLng: LanguageSchema.enum.en,
-				namespace: "translation",
-				supportedLngs: Languages,
+				...i18nSharedOptions,
 				backend: {
 					backends: [
 						HttpBackend, // if a namespace can't be loaded via normal http-backend loadPath, then the inMemoryLocalBackend will try to return the correct resources