+++ dev/vite-plugin-watch-workspace/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prosopo/vite-plugin-watch-workspace
 
+## 2.6.4
+### Patch Changes
+
+- Updated dependencies [745cc89]
+  - @prosopo/config@3.1.0
+
+## 2.6.3
+### Patch Changes
+
+- Updated dependencies [5619b4b]
+  - @prosopo/config@3.0.1
+
 ## 2.6.2
 ### Patch Changes