+++ packages/api/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @prosopo/api
 
+## 3.0.6
+### Patch Changes
+
+- b0d7207: Types for proper rotation
+- Updated dependencies [b0d7207]
+  - @prosopo/types@3.0.3
+  - @prosopo/user-access-policy@3.1.5
+
+## 3.0.5
+### Patch Changes
+
+  - @prosopo/user-access-policy@3.1.4
+
+## 3.0.4
+### Patch Changes
+
+  - @prosopo/user-access-policy@3.1.3
+
+## 3.0.3
+### Patch Changes
+
+- Updated dependencies [f682f0c]
+  - @prosopo/types@3.0.2
+  - @prosopo/user-access-policy@3.1.2
+
+## 3.0.2
+### Patch Changes
+
+  - @prosopo/types@3.0.1
+  - @prosopo/user-access-policy@3.1.1
+
+## 3.0.1
+### Patch Changes
+
+- Updated dependencies [913f2a6]
+  - @prosopo/user-access-policy@3.1.0
+
 ## 3.0.0
 ### Major Changes