+++ packages/file-server/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @prosopo/file-server
 
+## 2.6.4
+### Patch Changes
+
+  - @prosopo/util@3.0.2
+
+## 2.6.3
+### Patch Changes
+
+  - @prosopo/util@3.0.1
+
 ## 2.6.2
 ### Patch Changes