+++ packages/locale/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @prosopo/locale
 
+## 3.0.2
+### Patch Changes
+
+- f682f0c: Moving type and fixing i18n config
+
+## 3.0.1
+### Patch Changes
+
+- 87bd9bc: Allow non-explicit languages
+
 ## 3.0.0
 ### Major Changes