+++ packages/procaptcha-wrapper/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @prosopo/procaptcha-wrapper
 
+## 2.6.6
+### Patch Changes
+
+- Updated dependencies [b0d7207]
+  - @prosopo/types@3.0.3
+
+## 2.6.5
+### Patch Changes
+
+- Updated dependencies [f682f0c]
+  - @prosopo/locale@3.0.2
+  - @prosopo/types@3.0.2
+
+## 2.6.4
+### Patch Changes
+
+- Updated dependencies [87bd9bc]
+  - @prosopo/locale@3.0.1
+  - @prosopo/types@3.0.1
+
+## 2.6.3
+### Patch Changes
+
+- dead520: fix dep versions
+
+## 2.6.2
+### Patch Changes
+
+- 5705cfb: use build+bundle npm scripts
+
 ## 2.6.1
 ### Patch Changes