+++ packages/types/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @prosopo/types
 
+## 3.0.3
+### Patch Changes
+
+- b0d7207: Types for proper rotation
+
+## 3.0.2
+### Patch Changes
+
+- f682f0c: Moving type and fixing i18n config
+- Updated dependencies [f682f0c]
+  - @prosopo/locale@3.0.2
+  - @prosopo/common@3.0.2
+
+## 3.0.1
+### Patch Changes
+
+- Updated dependencies [87bd9bc]
+  - @prosopo/locale@3.0.1
+  - @prosopo/common@3.0.1
+
 ## 3.0.0
 ### Major Changes